requirements.txt
torch>=2.1.0
torchvision
transformers>=4.35.0
pillow
numpy
scikit-learn
tqdm